+++ src/screens/FamilyDang/FamilyInfo/dogprofile.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { DogProfile as Profile } from '~components/Profile/DogProfile';
+
+export const DogProfile = () => {
+  return <Profile dogId={1} />;
+};
